Fixed MinGW compiler warnings

+++ Emulator/Misc/RetroShell/Interpreter.cpp
@@ -318,11 +318,11 @@ Interpreter::help(const Command& current)
     usage(current);
 
     // Determine tabular positions to align the output
-    usize tab = 0;
+    isize tab = 0;
     for (auto &it : current.subCommands) {
-        tab = std::max(tab, it.fullName.length());
+        tab = std::max(tab, (isize)it.fullName.length());
     }
-    tab += indent.size();
+    tab += (isize)indent.size();
 
     isize group = -1;
